+++ Modules/Microsoft365DSC/DSCResources/MSFT_VivaEngagementRoleMember/MSFT_VivaEngagementRoleMember.psm1
@@ -6,25 +6,11 @@ function Get-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String]
-        $Identity,
+        $Role,
 
         [Parameter()]
-        [System.Boolean]
-        $DisableAudioVideoInCallsAndMeetings,
-
-        [Parameter()]
-        [System.Boolean]
-        $DisableCallsAndMeetings,
-
-        [Parameter()]
-        [ValidateSet('Enabled', 'Disabled')]
-        [System.String]
-        $VDI2Optimization,
-
-        [Parameter()]
-        [ValidateSet('Present', 'Absent')]
-        [System.String]
-        $Ensure = 'Present',
+        [System.String[]]
+        $Members,
 
         [Parameter()]
         [System.Management.Automation.PSCredential]
@@ -51,58 +37,58 @@ function Get-TargetResource
         $AccessTokens
     )
 
+    New-M365DSCConnection -Workload 'MicrosoftGraph' `
+        -InboundParameters $PSBoundParameters | Out-Null
+
+    #Ensure the proper dependencies are installed in the current environment.
+    Confirm-M365DSCDependencies
+
+    #region Telemetry
+    $ResourceName = $MyInvocation.MyCommand.ModuleName.Replace('MSFT_', '')
+    $CommandName = $MyInvocation.MyCommand
+    $data = Format-M365DSCTelemetryParameters -ResourceName $ResourceName `
+        -CommandName $CommandName `
+        -Parameters $PSBoundParameters
+    Add-M365DSCTelemetryEvent -Data $data
+    #endregion
+
+    $nullResult = $PSBoundParameters
     try
     {
-        if (-not $Script:exportedInstance -or $Script:exportedInstance.Identity -ne $Identity)
+        $uri = (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/employeeExperience/roles"
+        $roles = Invoke-MgGraphRequest -uri $uri -Method GET
+        $roleInstance = $roles.value | Where-Object -FilterScript {$_.displayName -eq $role}
+
+        if ([System.String]::IsNullOrEmpty($roleInstance))
         {
-            $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftTeams' `
-                -InboundParameters $PSBoundParameters | Out-Null
-
-            #Ensure the proper dependencies are installed in the current environment.
-            Confirm-M365DSCDependencies
-
-            #region Telemetry
-            $ResourceName = $MyInvocation.MyCommand.ModuleName.Replace('MSFT_', '')
-            $CommandName = $MyInvocation.MyCommand
-            $data = Format-M365DSCTelemetryParameters -ResourceName $ResourceName `
-                -CommandName $CommandName `
-                -Parameters $PSBoundParameters
-            Add-M365DSCTelemetryEvent -Data $data
-            #endregion
-
-            $nullResult = $PSBoundParameters
-            $nullResult.Ensure = 'Absent'
-
-            $instance = Get-CsTeamsVdiPolicy -Identity $Identity -ErrorAction SilentlyContinue
-        }
-        else
-        {
-            $instance = $Script:exportedInstance
+            throw "Could not find role instance with name {$role}"
         }
 
-        if ($null -eq $instance)
+        $uri = (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/employeeExperience/roles/$($roleInstance.id)/members"
+        $idsOfMembers = Invoke-MgGraphRequest -uri $uri -Method GET
+
+        $membersValue = @()
+        foreach ($memberId in $idsOfMembers.value)
         {
-            return $nullResult
+            $userInfo = Get-MgUser -UserId $memberId.id
+            $membersValue += $userInfo.UserPrincipalName
         }
 
-        Write-Verbose -Message "Found an instance with Identity {$Identity}"
         $results = @{
-            Identity                            = $instance.Identity
-            DisableAudioVideoInCallsAndMeetings = $instance.DisableAudioVideoInCallsAndMeetings
-            DisableCallsAndMeetings             = $instance.DisableCallsAndMeetings
-            VDI2Optimization                    = $instance.VDI2Optimization
-            Ensure                              = 'Present'
-            Credential                          = $Credential
-            ApplicationId                       = $ApplicationId
-            TenantId                            = $TenantId
-            CertificateThumbprint               = $CertificateThumbprint
-            ManagedIdentity                     = $ManagedIdentity.IsPresent
-            AccessTokens                        = $AccessTokens
+            Role                  = $Role
+            Members               = $membersValue
+            Credential            = $Credential
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
+            ManagedIdentity       = $ManagedIdentity.IsPresent
+            AccessTokens          = $AccessTokens
         }
         return [System.Collections.Hashtable] $results
     }
     catch
     {
+        Write-Verbose -Message $_
         New-M365DSCLogEntry -Message 'Error retrieving data:' `
             -Exception $_ `
             -Source $($MyInvocation.MyCommand.Source) `
@@ -120,25 +106,11 @@ function Set-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String]
-        $Identity,
+        $Role,
 
         [Parameter()]
-        [System.Boolean]
-        $DisableAudioVideoInCallsAndMeetings,
-
-        [Parameter()]
-        [System.Boolean]
-        $DisableCallsAndMeetings,
-
-        [Parameter()]
-        [ValidateSet('Enabled', 'Disabled')]
-        [System.String]
-        $VDI2Optimization,
-
-        [Parameter()]
-        [ValidateSet('Present', 'Absent')]
-        [System.String]
-        $Ensure = 'Present',
+        [System.String[]]
+        $Members,
 
         [Parameter()]
         [System.Management.Automation.PSCredential]
@@ -164,9 +136,6 @@ function Set-TargetResource
         [System.String[]]
         $AccessTokens
     )
-
-    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftTeams' `
-        -InboundParameters $PSBoundParameters | Out-Null
 
     #Ensure the proper dependencies are installed in the current environment.
     Confirm-M365DSCDependencies
@@ -182,58 +151,38 @@ function Set-TargetResource
 
     $currentInstance = Get-TargetResource @PSBoundParameters
 
-    $PSBoundParameters.Remove('Ensure') | Out-Null
-    $PSBoundParameters.Remove('Credential') | Out-Null
-    $PSBoundParameters.Remove('ApplicationId') | Out-Null
-    $PSBoundParameters.Remove('ApplicationSecret') | Out-Null
-    $PSBoundParameters.Remove('TenantId') | Out-Null
-    $PSBoundParameters.Remove('CertificateThumbprint') | Out-Null
-    $PSBoundParameters.Remove('ManagedIdentity') | Out-Null
-    $PSBoundParameters.Remove('AccessTokens') | Out-Null
+    $uri = (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/employeeExperience/roles"
+    $roles = Invoke-MgGraphRequest -uri $uri -Method GET
+    $roleInstance = $roles.value | Where-Object -FilterScript {$_.displayName -eq $role}
 
-    if ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Absent')
+    $membersDiff = Compare-Object -ReferenceObject $currentInstance.Members -DifferenceObject $Members
+
+    foreach ($member in $membersDiff)
     {
-        $CreateParameters = ([Hashtable]$PSBoundParameters).Clone()
-
-        $CreateParameters.Remove('Verbose') | Out-Null
-
-        $keys = $CreateParameters.Keys
-        foreach ($key in $keys)
+        $userInfo = Get-MgUser -Filter "UserPrincipalName eq '$($member.InputObject)'"
+        if (-not [System.String]::IsNullOrEmpty($userInfo))
         {
-            if ($null -ne $CreateParameters.$key -and $CreateParameters.$key.GetType().Name -like '*cimInstance*')
+            if ($member.SideIndicator -eq '=>')
             {
-                $keyValue = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $CreateParameters.$key
-                $CreateParameters.Remove($key) | Out-Null
-                $CreateParameters.Add($keyName, $keyValue)
+                Write-Verbose -Message "Adding user {$($member.InputObject)} to role {$Role}"
+                $uri = (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/employeeExperience/roles/$($roleInstance.id)/members"
+                $body = @{
+                    "user@odata.bind"= (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/users('" + $userInfo.Id + "')"
+                }
+                Write-Verbose -Message "POST request to $uri with:`r`n$(ConvertTo-Json $body -Depth 10)"
+                Invoke-MgGraphRequest -uri $uri -Method POST -Body $body
+            }
+            else
+            {
+                Write-Verbose -Message "Removing user {$($member.InputObject)} from role {$Role}"
+                $uri = (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/employeeExperience/roles/$($roleInstance.id)/members/$($userInfo.Id)"
+                Invoke-MgGraphRequest -uri $uri -Method DELETE
             }
         }
-        Write-Verbose -Message "Creating {$Identity} with Parameters:`r`n$(Convert-M365DscHashtableToString -Hashtable $CreateParameters)"
-        New-CsTeamsVdiPolicy @CreateParameters | Out-Null
-    }
-    elseif ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Present')
-    {
-        Write-Verbose -Message "Updating {$Identity}"
-
-        $UpdateParameters = ([Hashtable]$PSBoundParameters).Clone()
-        $UpdateParameters.Remove('Verbose') | Out-Null
-
-        $keys = $UpdateParameters.Keys
-        foreach ($key in $keys)
+        else
         {
-            if ($null -ne $UpdateParameters.$key -and $UpdateParameters.$key.GetType().Name -like '*cimInstance*')
-            {
-                $keyValue = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $UpdateParameters.$key
-                $UpdateParameters.Remove($key) | Out-Null
-                $UpdateParameters.Add($keyName, $keyValue)
-            }
+            Write-Error -Message "Could not find user {$($member.InputObject)}"
         }
-
-        Set-CsTeamsVdiPolicy @UpdateParameters | Out-Null
-    }
-    elseif ($Ensure -eq 'Absent' -and $currentInstance.Ensure -eq 'Present')
-    {
-        Write-Verbose -Message "Removing {$Identity}"
-        Remove-CsTeamsVdiPolicy -Identity $currentInstance.Identity
     }
 }
 
@@ -245,25 +194,11 @@ function Test-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String]
-        $Identity,
+        $Role,
 
         [Parameter()]
-        [System.Boolean]
-        $DisableAudioVideoInCallsAndMeetings,
-
-        [Parameter()]
-        [System.Boolean]
-        $DisableCallsAndMeetings,
-
-        [Parameter()]
-        [ValidateSet('Enabled', 'Disabled')]
-        [System.String]
-        $VDI2Optimization,
-
-        [Parameter()]
-        [ValidateSet('Present', 'Absent')]
-        [System.String]
-        $Ensure = 'Present',
+        [System.String[]]
+        $Members,
 
         [Parameter()]
         [System.Management.Automation.PSCredential]
@@ -302,26 +237,11 @@ function Test-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
-    Write-Verbose -Message "Testing configuration of {$Identity}"
-
     $CurrentValues = Get-TargetResource @PSBoundParameters
     $ValuesToCheck = ([Hashtable]$PSBoundParameters).Clone()
-    $ValuesToCheck.Remove('Identity') | Out-Null
-
-
 
     Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
     Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $ValuesToCheck)"
-
-    #Convert any DateTime to String
-    foreach ($key in $ValuesToCheck.Keys)
-    {
-        if (($null -ne $CurrentValues[$key]) `
-                -and ($CurrentValues[$key].GetType().Name -eq 'DateTime'))
-        {
-            $CurrentValues[$key] = $CurrentValues[$key].toString()
-        }
-    }
 
     $testResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
         -Source $($MyInvocation.MyCommand.Source) `
@@ -368,7 +288,7 @@ function Export-TargetResource
         $AccessTokens
     )
 
-    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftTeams' `
+    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' `
         -InboundParameters $PSBoundParameters
 
     #Ensure the proper dependencies are installed in the current environment.
@@ -385,34 +305,32 @@ function Export-TargetResource
 
     try
     {
-        [array]$getValue = Get-CsTeamsVdiPolicy -ErrorAction Stop
+        $Script:ExportMode = $true
+        $uri = (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/employeeExperience/roles"
+
+        [array] $Script:exportedInstances = Invoke-MgGraphRequest -Uri $uri -Method Get -ErrorAction Stop
 
         $i = 1
         $dscContent = ''
-        if ($getValue.Length -eq 0)
+        if ($Script:exportedInstances.Length -eq 0)
         {
-            Write-M365DSCHost -Message $Global:M365DSCEmojiGreenCheckMark -CommitWrite
+            Write-M365DSCHost -Message $Global:M365DSCEmojiGreenCheckMark
         }
         else
         {
             Write-M365DSCHost -Message "`r`n" -DeferWrite
         }
-        foreach ($config in $getValue)
+        foreach ($config in $Script:exportedInstances.value)
         {
             if ($null -ne $Global:M365DSCExportResourceInstancesCount)
             {
                 $Global:M365DSCExportResourceInstancesCount++
             }
 
-            $displayedKey = $config.Identity
-            if (-not [String]::IsNullOrEmpty($config.displayName))
-            {
-                $displayedKey = $config.displayName
-            }
-            Write-M365DSCHost -Message "    |---[$i/$($getValue.Count)] $displayedKey" -DeferWrite
+            $displayedKey = $config.displayName
+            Write-M365DSCHost -Message "    |---[$i/$($Script:exportedInstances.Count)] $displayedKey" -DeferWrite
             $params = @{
-                Identity              = $config.Identity
-                Ensure                = 'Present'
+                Role                  = $config.displayName
                 Credential            = $Credential
                 ApplicationId         = $ApplicationId
                 TenantId              = $TenantId
@@ -421,9 +339,7 @@ function Export-TargetResource
                 AccessTokens          = $AccessTokens
             }
 
-            $Script:exportedInstance = $config
             $Results = Get-TargetResource @Params
-
             $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
                 -ConnectionMode $ConnectionMode `
                 -ModulePath $PSScriptRoot `